Add minus characters to YAML multiline indicators
The minus characters indicate that no newline should be appended onto
the end of the multiline string.  In this case that is unwanted, since
the multiline string represents a username, password, or other secret
whose value is changed by appending a newline.

See here for more information about YAML multiline jazz:
https://yaml-multiline.info/